feat: app wrappers modularization
